requirements.txt
albumentations==1.3.0
numpy==1.24.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
pandas==2.0.0
Pillow==9.5.0
pyparsing==3.0.9
PyYAML==6.0
scikit-build==0.17.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
timm==0.6.13
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0